backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
python-multipart
python-dotenv
openai
anthropic
openai
pypdf
